+++ app/scenes/Document/components/History/EventListItem.tsx
@@ -12,11 +12,11 @@ import {
 import { useTranslation } from "react-i18next";
 import styled, { css } from "styled-components";
 import { s } from "@shared/styles";
+import Text from "@shared/components/Text";
 import type Document from "~/models/Document";
 import type Event from "~/models/Event";
 import Time from "~/components/Time";
 import Logger from "~/utils/Logger";
-import Text from "./Text";
 
 type Props = {
   document: Document;
