+++ shared/i18n/locales/en_US/translation.json
@@ -324,15 +324,6 @@
   "our engineers have been notified": "our engineers have been notified",
   "Clear cache + reload": "Clear cache + reload",
   "Show detail": "Show detail",
-  "{{userName}} archived": "{{userName}} archived",
-  "{{userName}} restored": "{{userName}} restored",
-  "{{userName}} deleted": "{{userName}} deleted",
-  "{{userName}} added {{addedUserName}}": "{{userName}} added {{addedUserName}}",
-  "{{userName}} removed {{removedUserName}}": "{{userName}} removed {{removedUserName}}",
-  "{{userName}} moved from trash": "{{userName}} moved from trash",
-  "{{userName}} published": "{{userName}} published",
-  "{{userName}} unpublished": "{{userName}} unpublished",
-  "{{userName}} moved": "{{userName}} moved",
   "A ZIP file containing the images, and documents in the Markdown format.": "A ZIP file containing the images, and documents in the Markdown format.",
   "A ZIP file containing the images, and documents as HTML files.": "A ZIP file containing the images, and documents as HTML files.",
   "Structured data that can be used to transfer data to another compatible {{ appName }} instance.": "Structured data that can be used to transfer data to another compatible {{ appName }} instance.",
@@ -424,12 +415,6 @@
   "{{ hours }}h {{ minutes }}m read": "{{ hours }}h {{ minutes }}m read",
   "{{ hours }}h read": "{{ hours }}h read",
   "{{ minutes }}m read": "{{ minutes }}m read",
-  "Revision deleted": "Revision deleted",
-  "{{count}} people": "{{count}} person",
-  "{{count}} people_plural": "{{count}} people",
-  "Current version": "Current version",
-  "{{userName}} edited": "{{userName}} edited",
-  "Revision options": "Revision options",
   "Manage": "Manage",
   "All members": "All members",
   "Everyone in the workspace": "Everyone in the workspace",
@@ -725,6 +710,7 @@
   "Revoking": "Revoking",
   "Are you sure you want to revoke access?": "Are you sure you want to revoke access?",
   "Delete app": "Delete app",
+  "Revision options": "Revision options",
   "Share options": "Share options",
   "Headings you add to the document will appear here": "Headings you add to the document will appear here",
   "Contents": "Contents",
@@ -832,10 +818,24 @@
   "Archived": "Archived",
   "Save draft": "Save draft",
   "Restore version": "Restore version",
-  "Previous revision": "Previous revision",
+  "{{userName}} archived": "{{userName}} archived",
+  "{{userName}} restored": "{{userName}} restored",
+  "{{userName}} deleted": "{{userName}} deleted",
+  "{{userName}} added {{addedUserName}}": "{{userName}} added {{addedUserName}}",
+  "{{userName}} removed {{removedUserName}}": "{{userName}} removed {{removedUserName}}",
+  "{{userName}} moved from trash": "{{userName}} moved from trash",
+  "{{userName}} published": "{{userName}} published",
+  "{{userName}} unpublished": "{{userName}} unpublished",
+  "{{userName}} moved": "{{userName}} moved",
+  "Previous version": "Previous version",
+  "{{userName}} edited": "{{userName}} edited",
   "Highlight changes": "Highlight changes",
   "Compare to": "Compare to",
   "No history yet": "No history yet",
+  "Revision deleted": "Revision deleted",
+  "{{count}} people": "{{count}} person",
+  "{{count}} people_plural": "{{count}} people",
+  "Current version": "Current version",
   "Source": "Source",
   "Created": "Created",
   "Imported from {{ source }}": "Imported from {{ source }}",
